Move keyword api endpoints under controllers
