+++ src/controllers/mod.rs
@@ -1,3 +1,5 @@
 // TODO: All endpoints would be moved to submodules here
 
 pub mod helpers;
+
+pub mod keyword;
